+++ SlaskEngine/sources/LogHandler.cpp
@@ -104,14 +104,14 @@ void LogHandler::writeLog(const char *logger,const char *str,int type)
 	}
 }
 
-void LogHandler::log(int i)
+void LogHandler::log(double i)
 {
 	std::ostringstream str;
 	str << i;
 	instance()->writeLog(0, str.str().c_str(), Log);
 }
 
-void LogHandler::log(double i)
+void LogHandler::log(int i)
 {
 	std::ostringstream str;
 	str << i;
